refactor(server): extract library batch routes to submodule
Moves the three /api/library/batch/* handlers (batch_add_to_playlist, batch_delete_items, batch_set_genre), their request structs, and the dedupe_positive_ids helper out of routes.rs into routes/library_batch_routes.rs. Verbatim move - no behavior change. routes.rs shrinks ~266 lines; 628 tests still pass.

dedupe_positive_ids was used only by these three handlers, so it moves with them as a private helper. Follows the existing routes/ submodule pattern (pub(super) handlers, explicit imports).
